Allow client to force server to terminate if it exits
This is intended to be used by screen lockers, where the server must
exit if the screen locker does, to avoid a security hole.  As suggested
by Alan Coopersmith, this is implemented using a new flag for the
disconnect mode of the XFixes extension.  Includes wording suggestions
from Peter Hutterer.

Signed-off-by: Demi Marie Obenour <demiobenour@gmail.com>