keysymdef.h: add dead_hamza symbol
It is intended to be added to the ara(mac-phonetic) and my(phonetic) layouts on
the <BKSL> key. They are based on the following layouts listed in the CLDR:

- https://github.com/unicode-org/cldr/blob/main/keyboards/osx/ar-t-k0-osx-qwerty.xml
- https://github.com/unicode-org/cldr/blob/main/keyboards/osx/ms-t-k0-osx.xml

In CLDR, the dead keys are defined in the <transforms> section. Hamza “ء” is
used there in the first position, making it a dead key. Now, among the keys
that produce this character, E02 has transform="no" while D13 has not. So hamza
on D13 is a dead key but it is not on E02.

The compose sequences required for the mentioned layouts are intended to be
added to libX11. Specifically, dead_hamza with alef, waw and yeh to generate
hamzaonalef, hamzaonwaw, and hamzaonyeh, respectively. Additional sequences
might be added as necessary.